Рефакторинг контекста БД
