+++ Tests/MailSender.ConsoleTest/Data/Artist.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.ComponentModel.DataAnnotations;
+
+namespace MailSender.ConsoleTest.Data
+{
+    public class Artist
+    {
+        public int Id { get; set; }
+
+        [Required, MinLength(3)]
+        public string Name { get; set; }
+
+        public DateTime Birthday { get; set; }
+    }
+}
