+++ Tests/MailSender.ConsoleTest/Data/SongsDB.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
+﻿using System.Collections.Generic;
 using System.Data.Entity;
 using System.Linq;
 using System.Text;
@@ -14,26 +11,5 @@ namespace MailSender.ConsoleTest.Data
         public DbSet<Track> Tracks { get; set; }
 
         public DbSet<Artist> Artists { get; set; }
-    }
-
-    [Table("Track")]
-    public class Track
-    {
-        public int Id { get; set; }
-
-        [Required]
-        public string Name { get; set; }
-
-        public int Length { get; set; }
-    }
-
-    public class Artist
-    {
-        public int Id { get; set; }
-
-        [Required, MinLength(3)]
-        public string Name { get; set; }
-
-        public DateTime Birthday { get; set; }
     }
 }
